Added service for smart meter logs by passing customerid

+++ Presentation/Nop.Web/Models/SmartMeterLogs/SmartMeterLogModel.cs
@@ -12,7 +12,7 @@ namespace Nop.Web.Models.SmartMeterLogs
         public Guid DeviceID { get; set; }
         public string Lattitude { get; set; }
         public string Longitude { get; set; }
-        public double Consumption { get; set; }
+        public int Consumption { get; set; }
         public int TimeInterval { get; set; }
         public DateTime TimeIntervalSetTime { get; set; }
         public bool IsActive { get; set; }
